fix(ios): make referral route static-export compatible
Closes #5964
